Store the selected "move from" correctly
A-H is stored as 0-7, but rows are inverted so that 0 is the top row and 7 is the bottom row. This makes it easier to compute locations in BoardState since it's stored from left to right, top to bottom

I used a lookup table to invert row since it's just easier than doing it in code.